moderation-service/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
transformers>=4.51.0   # ShieldGemma2ForImageClassification landed in 4.51
accelerate>=0.30.0
torch>=2.2.0
Pillow>=10.0.0
pydantic>=2.0.0